refactor cli_utilities
