+++ src/main/python/afp_cli/cli.py
@@ -21,44 +21,25 @@ Options:
 """
 
 from __future__ import print_function, absolute_import, division
-from datetime import datetime
 import getpass
 import os
-import random
-import socket
 import subprocess
 import sys
 import tempfile
 
 from docopt import docopt
 from afp_cli import AWSFederationClientCmd, aws_credentials_file
-import afp_cli.cli_functions as cli
+from .cli_functions import (format_aws_credentials,
+                            get_default_afp_server,
+                            get_aws_credentials,
+                            get_first_role,
+                            format_account_and_role_list,
+                            )
 from afp_cli.log import error, debug
 from .password_providers import get_password
 from . import log
 from .config import load_config
 
-
-def get_default_afp_server():
-    """Return the FQDN of the host that is called "afp"
-
-    This is done by resolving "afp" into (potentially multiple) IPs.
-    One of those IPs is randomly chosen, then a reverse-lookup is performed
-    on that IP to get its FQDN.
-    """
-    try:
-        addrinfos = socket.getaddrinfo("afp", 443,
-                                       socket.AF_INET, socket.SOCK_STREAM)
-    except Exception as exc:
-        error("Could not resolve hostname 'afp': %s" % exc)
-    addrinfo = random.choice(addrinfos)
-    afp_server_ip = addrinfo[4][0]
-
-    try:
-        return socket.gethostbyaddr(afp_server_ip)[0]
-    except Exception as exc:
-        error("DNS reverse lookup failed for IP %s: %s" % (
-            afp_server_ip, exc))
 
 RC_SCRIPT_TEMPLATE = """
 # Pretend to be an interactive, non-login shell
@@ -88,7 +69,7 @@ def start_subshell(aws_credentials, role, account):
     rc_script = tempfile.NamedTemporaryFile(mode='w')
     rc_script.write(RC_SCRIPT_TEMPLATE.format(role=role, account=account,
                                               valid_seconds=aws_credentials['AWS_VALID_SECONDS']))
-    rc_script.write(cli.format_aws_credentials(aws_credentials, prefix='export '))
+    rc_script.write(format_aws_credentials(aws_credentials, prefix='export '))
     rc_script.flush()
     subprocess.call(
         ["bash", "--rcfile", rc_script.name],
@@ -99,40 +80,13 @@ def start_subshell(aws_credentials, role, account):
 def start_subcmd(aws_credentials, role, account):
     batch_file = tempfile.NamedTemporaryFile(suffix=".bat", delete=False)
     batch_file.write(BATCH_FILE_TEMPLATE.format(role=role, account=account))
-    batch_file.write(cli.format_aws_credentials(aws_credentials, prefix='set '))
+    batch_file.write(format_aws_credentials(aws_credentials, prefix='set '))
     batch_file.flush()
     batch_file.close()
     subprocess.call(
         ["cmd", "/K", batch_file.name])
     print("Left AFP subcmd.")
     os.unlink(batch_file.name)
-
-
-def get_first_role(federation_client, account):
-    try:
-        accounts_and_roles = federation_client.get_account_and_role_list()
-    except Exception as exc:
-        error("Failed to get account list from AWS: %s" % exc)
-
-    try:
-        return sorted(accounts_and_roles[account])[0]
-    except KeyError:
-        error("%s is not a valid AWS account" % account)
-    except IndexError:
-        error("Could not find any role for account %s" % account)
-
-
-def get_aws_credentials(federation_client, account, role):
-    try:
-        aws_credentials = federation_client.get_aws_credentials(account, role)
-    except Exception as exc:
-        error("Failed to get credentials from AWS: %s" % exc)
-
-    aws_credentials['AWS_VALID_SECONDS'] = cli.get_valid_seconds(aws_credentials['AWS_EXPIRATION_DATE'],
-                                                                 datetime.utcnow())
-    aws_credentials['AWS_ACCOUNT_NAME'] = account
-    aws_credentials['AWS_ASSUMED_ROLE'] = role
-    return aws_credentials
 
 
 def main():
@@ -165,13 +119,13 @@ def main():
         aws_credentials = get_aws_credentials(federation_client, account, role)
 
         if arguments['--show']:
-            print(cli.format_aws_credentials(aws_credentials))
+            print(format_aws_credentials(aws_credentials))
 
         elif arguments['--export']:
             if os.name == "nt":
-                print(cli.format_aws_credentials(aws_credentials, prefix='set '))
+                print(format_aws_credentials(aws_credentials, prefix='set '))
             else:
-                print(cli.format_aws_credentials(aws_credentials, prefix='export '))
+                print(format_aws_credentials(aws_credentials, prefix='export '))
         elif arguments['--write']:
             aws_credentials_file.write(aws_credentials)
         else:
@@ -186,6 +140,6 @@ def main():
                 error("Failed to start subshell: %s" % exc)
     else:
         try:
-            print(cli.format_account_and_role_list(federation_client.get_account_and_role_list()))
+            print(format_account_and_role_list(federation_client.get_account_and_role_list()))
         except Exception as exc:
             error("Failed to get account list from AWS: %s" % exc)
